server/requirements.txt
Django>=3.0.0,<3.1.0
djangorestframework>=3.11.0,<3.12.0
psycopg2>=2.7.5,<2.8.0
Pillow>=5.3.0,<5.4.0
flake8>=3.8.3,<3.9.0
django-cors-headers>=3.4.0,<3.5.0
